+++ src/test/java/com/upgenix/runners/CukesRunner.java
@@ -21,7 +21,7 @@ import org.junit.runner.RunWith;
         features = "src/test/resources/features",
         glue = "com/upgenix/step_definitions",
         dryRun = false,
-        tags = ""
+        tags = "@UPGNX-451"
 )
 
 public class CukesRunner {
