--- src/test/java/com/upgenix/step_definitions/Login.java
@@ -1,41 +1,0 @@
-package com.upgenix.step_definitions;
-
-import com.upgenix.pages.LoginPage;
-import com.upgenix.utilities.ConfigurationReader;
-import com.upgenix.utilities.Driver;
-import io.cucumber.java.en.And;
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-import org.junit.Assert;
-
-public class Login {
-    LoginPage login =new LoginPage();
-    @Given("user is on the login page")
-    public void user_is_on_the_login_page() {
-        String url= ConfigurationReader.getProperty("url");
-        Driver.getDriver().get(url);
-    }
-    @When("user enters SalesManager username")
-    public void user_enters_sales_manager_username() {
-        login.username.sendKeys("salesmanager15@info.com");
-
-    }
-    @When("user enters SalesManager password")
-    public void user_enters_sales_manager_password() {
-        login.password.sendKeys("salesmanager");
-
-    }
-
-
-    @Then("user should see the dashboard")
-    public void user_should_see_the_dashboard() {
-        Assert.assertEquals("SalesManager15", login.usertext.getText());
-
-
-    }
-    @And("user click login button")
-    public void userClickLoginButton() {
-        login.loginBtn.click();
-    }
-}
